requirements.txt
attrs==23.1.0
Bottleneck==1.3.7
branca==0.6.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.11.0
decorator==5.1.1
earthengine-api==0.1.395
fiona==1.9.5
folium==0.14.0
fonttools==4.25.0
GDAL==3.6.2
geopandas==0.14.2
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.3
joblib==1.2.0
kiwisolver==1.4.4
mapclassify==2.5.0
MarkupSafe==2.1.3
matplotlib==3.8.0
munkres==1.1.4
networkx==3.1
numexpr==2.8.7
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pip==23.3.1
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.0.9
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retry==0.9.2
rsa==4.9
Rtree==1.0.1
scikit-learn==1.3.0
scipy==1.12.0
setuptools==68.2.2
shapely==2.0.1
six==1.16.0
threadpoolctl==2.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.41.2
xyzservices==2022.9.0
